requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
requests==2.31.0
python-dotenv==1.0.0
openpyxl==3.1.2
pytest==7.4.2